// deleghebowcl/src/app/model/dichiarazioneDettaglio.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/
import {Cittadino} from './cittadino';
import {DichiarazioneDettaglioStato} from './dichiarazioneDettaglioStato';
import {DichiarazioneRuolo} from './dichiarazioneRuolo';
import {Documento} from './documento';
import { Entita } from './entita';


export interface DichiarazioneDettaglio extends Entita {
  uuid?: string;
  motivazione_casella?: string;
  stato?: DichiarazioneDettaglioStato;
  genitore_tutore_curatore?: Cittadino;
  ruolo_genitore_tutore_curatore?: DichiarazioneRuolo;
  figlio_tutelato_curato?: Cittadino;
  ruolo_figlio_tutelato_curato?: DichiarazioneRuolo;
  documento?: Documento;
  dicdet_id?: number;
  doc_id?: number;
}