// deleghebowcl/src/app/model/documento.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/
import { DocumentoFile } from './documentoFile';
import { DocumentoTipo } from './documentoTipo';
import { Entita } from './entita';


export interface Documento extends Entita{
    id?: number;
    desc?: string;
    data_scadenza_doc?: Date;
    data_rilascio?: Date;
    doc_desc?: string;
    numero_documento?: string;
    autorita?: string;
    tipo?: DocumentoTipo;
    file?: DocumentoFile;
    file_id?: number;
}